requirements-dev.txt
aiocontextvars==0.2.2
aiodns==2.0.0
aiofiles==0.4.0
aiohttp==3.5.4
alabaster==0.7.12
alembic==1.0.10
argh==0.26.2
arrow==0.14.2
asn1crypto==0.24.0
astroid==2.2.5
async-generator==1.10
async-timeout==3.0.1
asyncpg==0.18.3
atomicwrites==1.3.0
attrs==19.1.0
Babel==2.7.0
binaryornot==0.4.4
cchardet==2.1.4
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
Click==7.0
contextvars==2.4
cookiecutter==1.6.0
cryptography==2.7
doc8==0.8.0
docutils==0.14
future==0.17.1
gidgethub==3.1.0
gino==0.8.3
h11==0.8.1
h2==3.1.0
hpack==3.0.0
httpcore==0.3.0
httptools==0.0.13
hyperframe==5.2.0
idna==2.8
idna-ssl==1.1.0
imagesize==1.1.0
immutables==0.9
importlib-metadata==0.18
isort==4.3.20
Jinja2==2.10.1
jinja2-time==0.2.0
lazy-object-proxy==1.4.1
livereload==2.6.1
Mako==1.0.12
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==7.0.0
multidict==4.5.2
packaging==19.0
pathtools==0.1.2
pbr==5.3.1
pluggy==0.12.0
port-for==0.3.1
poyo==0.4.2
psycopg2-binary==2.8.3
py==1.8.0
pycares==3.0.0
pycparser==2.19
Pygments==2.4.2
PyJWT==1.7.1
pylint==2.3.1
pyparsing==2.4.0
pytest==4.6.3
pytest-sanic==1.0.0
python-dateutil==2.8.0
python-editor==1.0.4
pytz==2019.1
PyYAML==5.1.1
requests==2.22.0
requests-async==0.5.0
restructuredtext-lint==1.3.0
rfc3986==1.3.2
sanic==19.6.0
sanic-jinja2==0.7.4
six==1.12.0
snowballstemmer==1.2.1
Sphinx==2.1.2
sphinx-autobuild==0.7.1
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
SQLAlchemy==1.3.5
stevedore==1.30.1
tornado==6.0.3
typed-ast==1.4.0
typing==3.6.6
typing-extensions==3.7.2
ujson==1.35
uritemplate==3.0.0
urllib3==1.25.3
uvloop==0.12.2
watchdog==0.9.0
wcwidth==0.1.7
websockets==6.0
whichcraft==0.5.2
wrapt==1.11.1
yarl==1.3.0
zipp==0.5.1
